+++ miden-crypto/src/dsa/falcon512_rpo/hash_to_point.rs
@@ -46,7 +46,7 @@ pub fn hash_to_point_rpo256(message: Word, nonce: &Nonce) -> Polynomial<FalconFe
 
 /// Returns a polynomial in Z_p[x]/(phi) representing the hash of the provided message and
 /// nonce using SHAKE256. This is the hash-to-point algorithm used in the reference implementation.
-#[cfg(test)]
+#[cfg(all(test, feature = "std"))]
 pub fn hash_to_point_shake256(message: &[u8], nonce: &Nonce) -> Polynomial<FalconFelt> {
     use sha3::{
         Shake256,
@@ -93,7 +93,7 @@ fn felt_to_falcon_felt(value: Felt) -> FalconFelt {
 /// Note that since `FalconFelt::new` accepts `i16`, we first reduce the `u32` value modulo
 /// the Falcon prime and then cast the resulting value to an `i16`.
 /// Note that this final cast is safe as the Falcon prime is less than `i16::MAX`.
-#[cfg(test)]
+#[cfg(all(test, feature = "std"))]
 fn u32_to_falcon_felt(value: u32) -> FalconFelt {
     FalconFelt::new((value % MODULUS as u32) as i16)
 }