+++ miden-crypto/src/dsa/falcon512_rpo/math/mod.rs
@@ -8,8 +8,6 @@
 use alloc::vec::Vec;
 use core::ops::MulAssign;
 
-#[cfg(not(feature = "std"))]
-use num::Float;
 use num::{BigInt, FromPrimitive, One, Zero};
 use num_complex::Complex64;
 use rand::Rng;