+++ miden-crypto/src/merkle/mmr/forest.rs
@@ -3,10 +3,11 @@ use core::{
     ops::{BitAnd, BitOr, BitXor, BitXorAssign},
 };
 
-use winter_utils::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable};
-
 use super::InOrderIndex;
-use crate::Felt;
+use crate::{
+    Felt,
+    utils::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable},
+};
 
 /// A compact representation of trees in a forest. Used in the Merkle forest (MMR).
 ///