+++ miden-crypto/src/merkle/smt/mod.rs
@@ -6,10 +6,12 @@ use core::{
     hash::Hash,
 };
 
-use winter_utils::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable};
-
 use super::{EmptySubtreeRoots, InnerNodeInfo, MerkleError, NodeIndex, SparseMerklePath};
-use crate::{EMPTY_WORD, Felt, Map, Word, hash::rpo::Rpo256};
+use crate::{
+    EMPTY_WORD, Felt, Map, Word,
+    hash::rpo::Rpo256,
+    utils::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable},
+};
 
 mod full;
 pub use full::{MAX_LEAF_ENTRIES, SMT_DEPTH, Smt, SmtLeaf, SmtLeafError, SmtProof, SmtProofError};